requirements.txt
streamlit
requests
PyYAML
python-dotenv
aiohttp
